ecom_base/pyproject.toml
[tool.poetry]
name = "ecom-base"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
pydantic = {extras = ["email"], version = "^2.10.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
